вынесет функции для работы со временем в отдельный модуль
