+++ src/mock/points.js
@@ -1,4 +1,5 @@
-import { getRandomArrayEl, getRandomRange, getRandomDate } from '../utils/utils.js';
+import { getRandomArrayEl, getRandomRange } from '../utils/utils.js';
+import { getRandomDate } from '../utils/date-utils.js';
 import { POINTS_TYPES } from '../utils/const.js';
 import { getOffersByType } from './offers.js';
 
@@ -7,8 +8,8 @@ const mockPoints = [
   {
     id: '00000000-0000-0000-0000-point0000000', //f4b62099-293f-4c3d-a702-94eec4a2808c
     basePrice: getRandomRange(10, 10000),
-    dateFrom: getRandomDate(new Date(2025, 0, 1), new Date(2026, 0, 1)),
-    dateTo: getRandomDate(new Date(2025, 0, 1), new Date(2026, 0, 2)),
+    dateFrom: getRandomDate(new Date(2025, 1, 1), new Date(2025, 1, 2)),
+    dateTo: getRandomDate(new Date(2025, 1, 3), new Date(2025, 1, 4)),
     destination: '00000000-0000-0000-0000-dest00000000', //bfa5cb75-a1fe-4b77-a83c-0e528e910e04
     isFavorite: !!getRandomRange(0,1),
     offers: getRandomOffersList(getOffersByType(POINTS_TYPES[0])),
@@ -17,8 +18,8 @@ const mockPoints = [
   {
     id: '00000000-0000-0000-0000-point0000001',
     basePrice: getRandomRange(10, 10000),
-    dateFrom: getRandomDate(new Date(2025, 0, 1), new Date(2026, 0, 1)),
-    dateTo: getRandomDate(new Date(2025, 0, 1), new Date(2026, 0, 2)),
+    dateFrom: getRandomDate(new Date(2025, 1, 1), new Date(2025, 1, 2)),
+    dateTo: getRandomDate(new Date(2025, 1, 3), new Date(2025, 1, 4)),
     destination: '00000000-0000-0000-0000-dest00000001',
     isFavorite: !!getRandomRange(0,1),
     offers: getRandomOffersList(getOffersByType(POINTS_TYPES[1])),
@@ -27,8 +28,8 @@ const mockPoints = [
   {
     id: '00000000-0000-0000-0000-point0000002',
     basePrice: getRandomRange(10, 10000),
-    dateFrom: getRandomDate(new Date(2025, 0, 1), new Date(2026, 0, 1)),
-    dateTo: getRandomDate(new Date(2025, 0, 1), new Date(2026, 0, 2)),
+    dateFrom: getRandomDate(new Date(2025, 1, 1), new Date(2025, 1, 2)),
+    dateTo: getRandomDate(new Date(2025, 1, 3), new Date(2025, 1, 4)),
     destination: '00000000-0000-0000-0000-dest00000002',
     isFavorite: !!getRandomRange(0,1),
     offers: getRandomOffersList(getOffersByType(POINTS_TYPES[4])),
@@ -37,8 +38,8 @@ const mockPoints = [
   {
     id: '00000000-0000-0000-0000-point0000003',
     basePrice: getRandomRange(10, 10000),
-    dateFrom: getRandomDate(new Date(2024, 0, 1), new Date(2024, 0, 1)),
-    dateTo: getRandomDate(new Date(2024, 0, 1), new Date(2024, 0, 2)),
+    dateFrom: getRandomDate(new Date(2025, 1, 1), new Date(2025, 1, 2)),
+    dateTo: getRandomDate(new Date(2025, 1, 3), new Date(2025, 1, 4)),
     destination: '00000000-0000-0000-0000-dest00000003',
     isFavorite: !!getRandomRange(0,1),
     offers: getRandomOffersList(getOffersByType(POINTS_TYPES[7])),
@@ -47,8 +48,8 @@ const mockPoints = [
   {
     id: '00000000-0000-0000-0000-point0000004',
     basePrice: getRandomRange(10, 10000),
-    dateFrom: getRandomDate(new Date(2024, 0, 1), new Date(2024, 0, 1)),
-    dateTo: getRandomDate(new Date(2024, 0, 1), new Date(2024, 0, 2)),
+    dateFrom: getRandomDate(new Date(2025, 1, 1), new Date(2025, 1, 2)),
+    dateTo: getRandomDate(new Date(2025, 1, 3), new Date(2025, 1, 4)),
     destination: '00000000-0000-0000-0000-dest00000004',
     isFavorite: !!getRandomRange(0,1),
     offers: getRandomOffersList(getOffersByType(POINTS_TYPES[3])),
@@ -57,8 +58,8 @@ const mockPoints = [
   {
     id: '00000000-0000-0000-0000-point0000005',
     basePrice: getRandomRange(10, 10000),
-    dateFrom: getRandomDate(new Date(2024, 0, 1), new Date(2024, 11, 1)),
-    dateTo: getRandomDate(new Date(2024, 0, 1), new Date(2024, 0, 2)),
+    dateFrom: getRandomDate(new Date(2025, 1, 1), new Date(2025, 1, 2)),
+    dateTo: getRandomDate(new Date(2025, 1, 3), new Date(2025, 1, 4)),
     destination: '00000000-0000-0000-0000-dest00000005',
     isFavorite: !!getRandomRange(0,1),
     offers: getRandomOffersList(getOffersByType(POINTS_TYPES[2])),
