+++ src/model/points-model.js
@@ -1,6 +1,6 @@
 import { getRandomPoint } from '../mock/points.js';
 
-const POINTS_COUNT = 10;
+const POINTS_COUNT = 5;
 
 export default class PointsModel {
   #points;
